Updated README.md
I feel this change is needed as it can cause confusion in many new users. Even as a relatively experienced developer, This made me think I had no option but to turn to the developer for help. It upset me very much when I was unable to run this amazing project. This pull request should help mitigate confusion when running the project.